Code organization with a main function that calls both login and scrap functions
